+++ plotnine/ggplot.py
@@ -1,12 +1,18 @@
+from __future__ import annotations
+
 import sys
+import typing
+from collections.abc import Sequence
 from copy import deepcopy
 from itertools import chain
 from pathlib import Path
 from types import SimpleNamespace as NS
+from typing import Any, Iterable, Union
 from warnings import warn
 
 import pandas as pd
 import matplotlib as mpl
+import matplotlib.figure
 import matplotlib.pyplot as plt
 import matplotlib.transforms as mtransforms
 from matplotlib.offsetbox import AnchoredOffsetbox
@@ -23,17 +29,24 @@ from .exceptions import PlotnineError, PlotnineWarning
 from .scales.scales import Scales
 from .coords import coord_cartesian
 from .guides.guides import guides
-from .geoms import geom_blank
+
+# mypy believes there is a duplicate definition
+# of geom_blank even though it only appears once
+from .geoms import geom_blank  # type: ignore[no-redef]  # mypy bug
+
 from .utils import (
     defaults,
     from_inches,
     is_data_like,
     order_as_data_mapping,
     to_inches,
-    to_pandas,
     ungroup
 )
 
+if typing.TYPE_CHECKING:
+    import plotnine as p9
+
+    from .typing import DataLike, PlotAddable
 
 # Show plots if in interactive mode
 if sys.flags.interactive:
@@ -59,7 +72,12 @@ class ggplot:
         in which `ggplot()` is called.
     """
 
-    def __init__(self, data=None, mapping=None, environment=None):
+    def __init__(
+        self,
+        data: DataLike | None = None,
+        mapping: aes | None = None,
+        environment: dict[str, Any] | None = None
+    ) -> None:
         # Allow some sloppiness
         data, mapping = order_as_data_mapping(data, mapping)
         if mapping is None:
@@ -68,7 +86,7 @@ class ggplot:
         # Recognize plydata groups
         if hasattr(data, 'group_indices') and 'group' not in mapping:
             mapping = mapping.copy()
-            mapping['group'] = data.group_indices()
+            mapping['group'] = data.group_indices()  # type: ignore
 
         self.data = data
         self.mapping = mapping
@@ -80,12 +98,15 @@ class ggplot:
         self.theme = theme_get()
         self.coordinates = coord_cartesian()
         self.environment = environment or EvalEnvironment.capture(1)
-        self.layout = None
-        self.figure = None
-        self.watermarks = []
+        self.layout = Layout()
+        self.figure: mpl.figure.Figure | None = None
+        self.watermarks: list[p9.watermark] = []
         self.axs = None
 
-    def __str__(self):
+        # build artefacts
+        self._build_objs = NS()
+
+    def __str__(self) -> str:
         """
         Print/show the plot
         """
@@ -94,14 +115,14 @@ class ggplot:
         # Return and empty string so that print(p) is "pretty"
         return ''
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         """
         Print/show the plot
         """
         self.__str__()
         return '<ggplot: (%d)>' % self.__hash__()
 
-    def __deepcopy__(self, memo):
+    def __deepcopy__(self, memo: dict[Any, Any]) -> ggplot:
         """
         Deep copy without copying the dataframe and environment
         """
@@ -112,7 +133,7 @@ class ggplot:
         new = result.__dict__
 
         # don't make a deepcopy of data, or environment
-        shallow = {'data', 'environment', 'figure'}
+        shallow = {'data', 'environment', 'figure', '_build_objs'}
         for key, item in old.items():
             if key in shallow:
                 new[key] = old[key]
@@ -122,47 +143,49 @@ class ggplot:
 
         return result
 
-    def __iadd__(self, other):
+    def __iadd__(
+        self,
+        other: PlotAddable | list[PlotAddable] | None
+    ) -> ggplot:
         """
         Add other to ggplot object
-        """
-        try:
-            return other.__radd__(self, inplace=True)
-        except TypeError:
-            return other.__radd__(self)
-        except AttributeError as err:
-            if other is None:
-                return self.__add__(other)
-            raise err
-
-    def __add__(self, other):
-        """
-        Add to ggitems from a list
 
         Parameters
         ----------
-        other : object or list
+        other : object or Sequence
             Either an object that knows how to "radd"
             itself to a ggplot, or a list of such objects.
         """
-        if isinstance(other, list):
-            self = deepcopy(self)
+        if isinstance(other, Sequence):
             for item in other:
-                self += item
+                item.__radd__(self)
             return self
         elif other is None:
-            return deepcopy(self)
+            return self
         else:
             return other.__radd__(self)
 
-    def __rrshift__(self, other):
+    def __add__(self, other: PlotAddable | list[PlotAddable] | None) -> ggplot:
+        """
+        Add to ggplot from a list
+
+        Parameters
+        ----------
+        other : object or Sequence
+            Either an object that knows how to "radd"
+            itself to a ggplot, or a list of such objects.
+        """
+        self = deepcopy(self)
+        return self.__iadd__(other)
+
+    def __rrshift__(self, other: DataLike) -> ggplot:
         """
         Overload the >> operator to receive a dataframe
         """
         other = ungroup(other)
         if is_data_like(other):
             if self.data is None:
-                self.data = to_pandas(other)
+                self.data = other
             else:
                 raise PlotnineError(
                     "`>>` failed, ggplot object has data."
@@ -172,14 +195,12 @@ class ggplot:
             raise TypeError(msg.format(type(other)))
         return self
 
-    def draw(self, return_ggplot=False, show=False):
+    def draw(self, show: bool = False) -> mpl.figure.Figure:
         """
         Render the complete plot
 
         Parameters
         ----------
-        return_ggplot : bool
-            If ``True``, return ggplot object.
         show : bool (default: False)
             Whether to show the plot.
 
@@ -187,20 +208,12 @@ class ggplot:
         -------
         fig : ~matplotlib.figure.Figure
             Matplotlib figure
-        plot : ggplot (optional)
-            The ggplot object used for drawn, if ``return_ggplot`` is
-            ``True``.
-
-        Notes
-        -----
-        This method does not modify the original ggplot object. You can
-        get the modified ggplot object with :py:`return_ggplot=True`.
         """
         # Do not draw if drawn already.
         # This prevents a needless error when reusing
         # figure & axes in the jupyter notebook.
         if self.figure:
-            return (self.figure, self) if return_ggplot else self.figure
+            return self.figure
 
         # Prevent against any modifications to the users
         # ggplot object. Do the copy here as we may/may not
@@ -212,7 +225,7 @@ class ggplot:
             # setup
             figure, axs = self._create_figure()
             self._setup_parameters()
-            self.facet.strips.generate()
+            self.facet.strips.generate()  # type: ignore[attr-defined]
             self._resize_panels()
 
             # Drawing
@@ -227,11 +240,7 @@ class ggplot:
             # Artist object theming
             self.theme.apply(figure, axs)
 
-            if return_ggplot:
-                output = self.figure, self
-            else:
-                output = self.figure
-        return output
+        return self.figure
 
     def _draw_using_figure(self, figure, axs):
         """
@@ -275,10 +284,9 @@ class ggplot:
         if not self.layers:
             self += geom_blank()
 
-        self.layout = Layout()
-        layers = self.layers
-        scales = self.scales
-        layout = self.layout
+        layers = self._build_objs.layers = self.layers
+        scales = self._build_objs.scales = self.scales
+        layout = self._build_objs.layout = self.layout
 
         # Update the label information for the plot
         layers.update_labels(self)
@@ -619,7 +627,7 @@ class ggplot:
         self.theme.apply_axs(self.axs)
         self.theme.apply_figure(self.figure)
 
-    def _save_filename(self, ext):
+    def _save_filename(self, ext: str) -> Path:
         """
         Make a filename for use by the save method
 
@@ -646,9 +654,19 @@ class ggplot:
         new_labels = defaults(mapping, default)
         self.labels = defaults(self.labels, new_labels)
 
-    def save(self, filename=None, format=None, path=None,
-             width=None, height=None, units='in',
-             dpi=None, limitsize=True, verbose=True, **kwargs):
+    def save(
+        self,
+        filename: Union[str, Path] | None = None,
+        format: str | None = None,
+        path: str | None = None,
+        width: float | None = None,
+        height: float | None = None,
+        units: str = 'in',
+        dpi: float | None = None,
+        limitsize: bool = True,
+        verbose: bool = True,
+        **kwargs: Any
+    ) -> None:
         """
         Save a ggplot object as an image file
 
@@ -728,22 +746,20 @@ class ggplot:
         if dpi is not None:
             self.theme = self.theme + theme(dpi=dpi)
 
-        fig, p = self.draw(return_ggplot=True)
-        with plot_context(p):
-            fig.savefig(filename, **fig_kwargs)
-            plt.close(fig)
+        fig = self.draw(show=False)
+        fig.savefig(filename, **fig_kwargs)
 
 
-def ggsave(plot, *arg, **kwargs):
-    """
-    Save a ggplot object as an image file
-
-    Use :meth:`ggplot.save` instead
-    """
-    return plot.save(*arg, **kwargs)
+ggsave = ggplot.save
 
 
-def save_as_pdf_pages(plots, filename=None, path=None, verbose=True, **kwargs):
+def save_as_pdf_pages(
+    plots: Iterable[ggplot],
+    filename: str | None = None,
+    path: str | None = None,
+    verbose: bool = True,
+    **kwargs: Any
+) -> None:
     """
     Save multiple :class:`ggplot` objects to a PDF file, one per page.
 
@@ -810,6 +826,7 @@ def save_as_pdf_pages(plots, filename=None, path=None, verbose=True, **kwargs):
     plots = iter(plots)
 
     # filename, depends on the object
+    filename: str | Path | None = filename  # broaden allowed type for var
     if filename is None:
         # Take the first element from the iterator, store it, and
         # use it to generate a file name
@@ -830,10 +847,6 @@ def save_as_pdf_pages(plots, filename=None, path=None, verbose=True, **kwargs):
             # Save as a page in the PDF file
             pdf.savefig(fig, **fig_kwargs)
 
-            # To conserve memory when plotting a large number of pages,
-            # close the figure whether or not there was an exception
-            plt.close(fig)
-
 
 class plot_context:
     """
@@ -848,11 +861,11 @@ class plot_context:
         exits.
     """
 
-    def __init__(self, plot, show=False):
+    def __init__(self, plot: ggplot, show: bool = False) -> None:
         self.plot = plot
         self.show = show
 
-    def __enter__(self):
+    def __enter__(self) -> plot_context:
         """
         Enclose in matplolib & pandas environments
         """
@@ -874,6 +887,8 @@ class plot_context:
         if exc_type is None:
             if self.show:
                 plt.show()
+            else:
+                plt.close(self.plot.figure)
         else:
             # There is an exception, close any figure
             if self.plot.figure is not None: