+++ doc/changelog.rst
@@ -5,6 +5,17 @@ v0.10.2
 -------
 (Not yet released)
 
+API Changes
+***********
+
+- :method:`~plotnine.ggplot.add` and :method:`~plotnine.ggplot.iadd` now
+  accept the same types of objects. In this change
+  :method:`~plotnine.ggplot.iadd` has gained the ability to accept a list
+  of objects.
+
+- :method:`~plotnine.ggplot.draw` no longer accepts the argument
+  ``return_ggplot`` and the return value is always a matplolib figure.
+
 Bug Fixes
 *********